requirements.txt
wmi
google-auth
google-auth-oauthlib
google-api-python-client